backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
google-genai
google-adk

